backend/requirements/base.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
django-cors-headers==4.6.0
django-filter==24.3
django-simple-history==3.7.0
django-treebeard==4.7.1
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
Pillow==11.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
openpyxl==3.1.5

# EDO Module Dependencies
beautifulsoup4==4.14.3
cffi==2.0.0
charset-normalizer==3.4.1
cryptography==46.0.3
docxtpl==0.20.2
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
pdfminer.six==20251230
pdfplumber==0.11.9
playwright==1.49.1
pycparser==2.23
pypdfium2==5.3.0
python-docx==1.2.0
reportlab==4.4.9
soupsieve==2.8.2
sqlparse==0.5.5
typing_extensions==4.15.0
